requirements.txt
streamlit
openai
groq
python-dotenv
